#Refact: Implementado classe de configuração exclusiva ConnectionFactoryConfig
