+++ src/main/java/br/com/ccs/messagedispatcher/config/ConnectionFactoryConfig.java
@@ -1,0 +1,37 @@
+package br.com.ccs.messagedispatcher.config;
+
+import br.com.ccs.messagedispatcher.config.properties.MessageDispatcherProperties;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ConnectionFactoryConfig {
+
+    private final Logger log = LoggerFactory.getLogger(ConnectionFactoryConfig.class);
+    private final MessageDispatcherProperties properties;
+
+    public ConnectionFactoryConfig(MessageDispatcherProperties properties) {
+        this.properties = properties;
+    }
+
+    @Bean
+    @SuppressWarnings("unused")
+    protected ConnectionFactory connectionFactory() {
+        log.debug("Configurando ConnectionFactory");
+        CachingConnectionFactory connectionFactory = new CachingConnectionFactory();
+        connectionFactory.setHost(properties.getHost());
+        connectionFactory.setPort(properties.getPort());
+        connectionFactory.setUsername(properties.getUsername());
+        connectionFactory.setPassword(properties.getPassword());
+        connectionFactory.setVirtualHost(properties.getVirtualHost());
+        connectionFactory.setPublisherConfirmType(CachingConnectionFactory.ConfirmType.CORRELATED);
+        connectionFactory.setPublisherReturns(true);
+
+        log.debug("ConnectionFactory configurada");
+        return connectionFactory;
+    }
+}
